requirements.txt
pyinstaller~=4.3
yfinance~=0.1.59
pytz~=2021.1
matplotlib~=3.4.1
mplfinance~=0.12.7a12
bs4~=0.0.1
ttkthemes~=3.2.2
requests_futures
tqdm~=4.60.0
pandas~=1.2.4
requests~=2.25.1
python-dateutil~=2.8.1
beautifulsoup4~=4.9.3
future~=0.18.2
numpy~=1.20.2
StringGenerator~=0.4.4